ch: refactor ranks
 - refactor ranks

[Finishes #173749690]
