+++ frontend/components/icons/cards/normal/ranks/index.ts
@@ -1,0 +1,13 @@
+export * from './ace';
+export * from './two';
+export * from './three';
+export * from './four';
+export * from './five';
+export * from './six';
+export * from './seven';
+export * from './eight';
+export * from './nine';
+export * from './ten';
+export * from './jack';
+export * from './queen';
+export * from './king';
